+++ src/novedades/index.ts
@@ -2,5 +2,6 @@
 export * from './types'
 export * from './novedades.repository'
 export * from './novedad.detector'
+export * from './media-chat.repository'
 export * from './novedades.service'
 export * from './admin.handler'